Adição do campo de submissão unica em modalidade

+++ app/Http/Controllers/Submissao/TrabalhoController.php
@@ -56,7 +56,7 @@ class TrabalhoController extends Controller
         $modalidade = Modalidade::find($idModalidade);
 
         $mytime = Carbon::now('America/Recife');
-        if (! $modalidade->estaEmPeriodoDeSubmissao()) {
+        if (!$modalidade->estaEmPeriodoDeSubmissao()) {
             $this->authorize('isCoordenadorOrCoordenadorDaComissaoCientifica', $evento);
         }
         // dd($formSubTraba);
@@ -101,9 +101,9 @@ class TrabalhoController extends Controller
     {
         $originalName = $file->getClientOriginalName();
         $extension = $file->getClientOriginalExtension();
-        $originalName = str_replace('.'.$extension, '', $originalName);
+        $originalName = str_replace('.' . $extension, '', $originalName);
         $originalName = $this->removerCaracteresEspeciais($originalName);
-        $path = $file->storeAs($path, $originalName.'.'.$extension);
+        $path = $file->storeAs($path, $originalName . '.' . $extension);
 
         return $path;
     }
@@ -164,7 +164,7 @@ class TrabalhoController extends Controller
           Verifique no formulário, quais os tipos permitidos.'])->withInput($validatedData);
             }
 
-            if ($modalidade->apresentacao && ! $request->tipo_apresentacao) {
+            if ($modalidade->apresentacao && !$request->tipo_apresentacao) {
                 return redirect()->back()->withErrors(['tipo_apresentacao' => 'Selecione a forma de apresentação do trabalho.'])->withInput($validatedData);
             }
 
@@ -184,7 +184,7 @@ class TrabalhoController extends Controller
 
             $trabalhosDoAutor = Trabalho::where('eventoId', $request->eventoId)->where('autorId', Auth::user()->id)->where('status', '!=', 'arquivado')->count();
             // $areaModalidade = AreaModalidade::where('areaId', $request->araeaId)->where('modalidadeId', $request->modalidadeId)->first();
-            Log::debug('Numero de trabalhos'.$evento);
+            Log::debug('Numero de trabalhos' . $evento);
             if ($evento->numMaxTrabalhos != null && $trabalhosDoAutor >= $evento->numMaxTrabalhos) {
                 return redirect()->back()->withErrors(['numeroMax' => 'Número máximo de trabalhos permitidos atingido.'])->withInput($validatedData);
             }
@@ -389,7 +389,7 @@ class TrabalhoController extends Controller
             return redirect()->route('evento.visualizar', ['id' => $request->eventoId])
                 ->with(['message' => 'Submissão concluída com sucesso!', 'class' => 'success']);
         } catch (\Throwable $th) {
-            Log::info('message'.$th->getMessage());
+            Log::info('message' . $th->getMessage());
 
             return redirect()->back()->with(['message' => 'Submissão não foi concluída!', 'class' => 'danger']);
         }
@@ -471,29 +471,31 @@ class TrabalhoController extends Controller
      */
     public function update(TrabalhoUpdateRequest $request, $id)
     {
+        // dd($request->all());
         $validatedData = $request->validated();
 
         $trabalho = Trabalho::find($id);
         $evento = $trabalho->evento;
 
-        $arquivo = $request->file('arquivo'.$id);
+        $arquivo = $request->file('arquivo' . $id);
+
         if ($arquivo != null && $this->validarTipoDoArquivo($arquivo, $trabalho->modalidade)) {
-            return redirect()->back()->withErrors(['arquivo'.$id => 'Extensão de arquivo enviado é diferente do permitido.
+            return redirect()->back()->withErrors(['arquivo' . $id => 'Extensão de arquivo enviado é diferente do permitido.
                                                                   Verifique no formulário, quais os tipos permitidos.'])->withInput($validatedData);
         }
 
-        $trabalho->titulo = $request->input('nomeTrabalho'.$id);
-        $trabalho->resumo = $request->input('resumo'.$id);
-        $trabalho->areaId = $request->input('area'.$id);
-        if ($request->input('modalidade'.$id) != $trabalho->modalidadeId && $trabalho->avaliado == 'Avaliado') {
-            return redirect()->back()->withErrors(['modalidadeError'.$id => 'Não é possível alterar a modalidade de um trabalho avaliado.'])->withInput($validatedData);
-        } elseif ($request->input('modalidade'.$id) != $trabalho->modalidadeId && $trabalho->atribuicoes->count() > 0) {
-            return redirect()->back()->withErrors(['modalidadeError'.$id => 'Não é possível alterar a modalidade de um trabalho com revisores atribuídos.'])->withInput($validatedData);
+        $trabalho->titulo = $request->input('nomeTrabalho' . $id);
+        $trabalho->resumo = $request->input('resumo' . $id);
+        $trabalho->areaId = $request->input('area' . $id);
+        if ($request->input('modalidade' . $id) != $trabalho->modalidadeId && $trabalho->avaliado == 'Avaliado') {
+            return redirect()->back()->withErrors(['modalidadeError' . $id => 'Não é possível alterar a modalidade de um trabalho avaliado.'])->withInput($validatedData);
+        } elseif ($request->input('modalidade' . $id) != $trabalho->modalidadeId && $trabalho->atribuicoes->count() > 0) {
+            return redirect()->back()->withErrors(['modalidadeError' . $id => 'Não é possível alterar a modalidade de um trabalho com revisores atribuídos.'])->withInput($validatedData);
         } else {
-            $trabalho->modalidadeId = $request->input('modalidade'.$id);
+            $trabalho->modalidadeId = $request->input('modalidade' . $id);
         }
 
-        if ($trabalho->modalidade->apresentacao && ! $request->tipo_apresentacao) {
+        if ($trabalho->modalidade->apresentacao && !$request->tipo_apresentacao) {
             return redirect()->back()->withErrors(['tipo_apresentacao' => 'Selecione a forma de apresentação do trabalho.'])->withInput($validatedData);
         }
 
@@ -504,7 +506,7 @@ class TrabalhoController extends Controller
 
         $coautoresExcluidos = collect();
         // TODO Checando a mudança do autor do trabalho, a inclusão e exclusão de coautores
-        foreach ($request->input('emailCoautor_'.$id) as $i => $email) {
+        foreach ($request->input('emailCoautor_' . $id) as $i => $email) {
             $usuario_do_coautor = User::where('email', $email)->first();
             // Chegando se existe do usuário cadastrado no sistema
             // Se existir é checado se o usario já é coautor do trabalho e adicionado nos coautores
@@ -539,7 +541,7 @@ class TrabalhoController extends Controller
                     'email' => $email,
                     'password' => bcrypt($passwordTemporario),
                     'usuarioTemp' => true,
-                    'name' => $request->input('nomeCoautor_'.$id)[$i],
+                    'name' => $request->input('nomeCoautor_' . $id)[$i],
                 ]);
 
                 if ($i != 0) {
@@ -573,13 +575,13 @@ class TrabalhoController extends Controller
 
         $coautoresExcluidos = $usuarios_dos_coautores->diff($coautoresExcluidos);
         foreach ($coautoresExcluidos as $usuario_do_coautor) {
-            if (! (in_array($usuario_do_coautor->email, $request->input('emailCoautor_'.$id)))) {
+            if (!(in_array($usuario_do_coautor->email, $request->input('emailCoautor_' . $id)))) {
                 $usuario_do_coautor->coautor->trabalhos()->detach($id);
             }
         }
 
         // atualizando a ordem dos coautores
-        $email_dos_coautores = $validatedData['emailCoautor_'.$id];
+        $email_dos_coautores = $validatedData['emailCoautor_' . $id];
         unset($email_dos_coautores[0]);
         foreach ($email_dos_coautores as $ordem => $email) {
             $id_autor = User::where('email', $email)->get()->first()->id;
@@ -626,39 +628,41 @@ class TrabalhoController extends Controller
             $trabalho->tipo_apresentacao = $request->tipo_apresentacao;
         }
 
-        if ($request->file('arquivo'.$id) != null) {
-            $path = "trabalhos/{$evento->id}/{$trabalho->id}";
-            $file = $request->file('arquivo'.$id);
-            $path = $this->salvarArquivoComNomeOriginal($file, $path);
+        if ($request->file('arquivo' . $id) != null) {
+            if ($trabalho->modalidade->submissaoUnica == false) {
+                $path = "trabalhos/{$evento->id}/{$trabalho->id}";
+                $file = $request->file('arquivo' . $id);
+                $path = $this->salvarArquivoComNomeOriginal($file, $path);
 
-            //É necessário excluir o arquivo da tabela de arquivo também ao editar um trabalho
-            //Não só fazer o Storage::delete() do arquivo
-            $arquivosTrabalho = $trabalho->arquivo()->where('versaoFinal', true)->get();
-            foreach ($arquivosTrabalho as $arquivoTrabalho) {
-                if (Storage::disk()->exists($arquivoTrabalho->nome)) {
-                    Storage::delete($arquivoTrabalho->nome);
+                //É necessário excluir o arquivo da tabela de arquivo também ao editar um trabalho
+                //Não só fazer o Storage::delete() do arquivo
+                $arquivosTrabalho = $trabalho->arquivo()->where('versaoFinal', true)->get();
+                foreach ($arquivosTrabalho as $arquivoTrabalho) {
+                    if (Storage::disk()->exists($arquivoTrabalho->nome)) {
+                        Storage::delete($arquivoTrabalho->nome);
+                    }
+                    $arquivoTrabalho->delete();
                 }
-                $arquivoTrabalho->delete();
+
+                $arquivo = Arquivo::create([
+                    'nome' => $path,
+                    'trabalhoId' => $trabalho->id,
+                    'versaoFinal' => true,
+                ]);
+
+                /*$arquivoAtual = $trabalho->arquivo()->where('versaoFinal', true)->first();
+                if (Storage::disk()->exists($arquivoAtual->nome)) {
+                  Storage::delete($arquivoAtual->nome);
+                  $arquivoAtual->delete();
+                }*/
             }
-
-            $arquivo = Arquivo::create([
-                'nome' => $path,
-                'trabalhoId' => $trabalho->id,
-                'versaoFinal' => true,
-            ]);
-
-            /*$arquivoAtual = $trabalho->arquivo()->where('versaoFinal', true)->first();
-            if (Storage::disk()->exists($arquivoAtual->nome)) {
-              Storage::delete($arquivoAtual->nome);
-              $arquivoAtual->delete();
-            }*/
         }
 
         if ($trabalho->modalidade->midiasExtra) {
             foreach ($trabalho->modalidade->midiasExtra as $midia) {
                 if ($request[$midia->hyphenizeNome()]) {
                     $consulta = $trabalho->midiasExtra()->where('midia_extra_id', $midia->id);
-                    if (! $consulta->exists()) {
+                    if (!$consulta->exists()) {
                         $trabalho->midiasExtra()->attach($midia->id);
                     }
                     $documento = $consulta->first()->pivot;
@@ -710,7 +714,7 @@ class TrabalhoController extends Controller
 
         $trabalho->update();
 
-        return redirect()->back()->with(['mensagem' => $trabalho->titulo.' editado com sucesso!']);
+        return redirect()->back()->with(['mensagem' => $trabalho->titulo . ' editado com sucesso!']);
     }
 
     /**
@@ -723,7 +727,7 @@ class TrabalhoController extends Controller
     {
         $trabalho = Trabalho::find($id);
         $agora = Carbon::now();
-        if (auth()->user()->id != $trabalho->autorId || ! $trabalho->modalidade->estaEmPeriodoDeSubmissao()) {
+        if (auth()->user()->id != $trabalho->autorId || !$trabalho->modalidade->estaEmPeriodoDeSubmissao()) {
             return abort(403);
         }
 
@@ -769,7 +773,7 @@ class TrabalhoController extends Controller
             'trabalhoId' => ['required', 'integer'],
         ]);
 
-        if (! $modalidade->estaEmPeriodoDeSubmissao()) {
+        if (!$modalidade->estaEmPeriodoDeSubmissao()) {
             return redirect()->back()->withErrors(['error' => 'O periodo de submissão para esse trabalho se encerrou.']);
         }
 
@@ -884,14 +888,16 @@ class TrabalhoController extends Controller
         $usuariosDaComissaoOrganizadora = $evento->usuariosDaComissaoOrganizadora;
         $usuarioLogado = auth()->user();
 
-        if ($evento->coordenadorId == $usuarioLogado->id
-        || $usuariosDaComissaoCientifica->contains($usuarioLogado)
-        || $usuariosDaComissaoOrganizadora->contains($usuarioLogado)
-        || $evento->userIsCoordComissaoCientifica($usuarioLogado)
-        || $evento->userIsCoordComissaoOrganizadora($usuarioLogado)
-        || $trabalho->autorId == $usuarioLogado->id
-        || $trabalhosCoautor->contains($trabalho->id)
-        || $usuarioLogado->administradors()->exists()) {
+        if (
+            $evento->coordenadorId == $usuarioLogado->id
+            || $usuariosDaComissaoCientifica->contains($usuarioLogado)
+            || $usuariosDaComissaoOrganizadora->contains($usuarioLogado)
+            || $evento->userIsCoordComissaoCientifica($usuarioLogado)
+            || $evento->userIsCoordComissaoOrganizadora($usuarioLogado)
+            || $trabalho->autorId == $usuarioLogado->id
+            || $trabalhosCoautor->contains($trabalho->id)
+            || $usuarioLogado->administradors()->exists()
+        ) {
             // dd($arquivo);
             if ($midia != null && Storage::disk()->exists($midia->caminho)) {
                 return Storage::download($midia->caminho);
@@ -947,14 +953,16 @@ class TrabalhoController extends Controller
         $usuariosDaComissaoOrganizadora = $evento->usuariosDaComissaoOrganizadora;
         $usuarioLogado = auth()->user();
 
-        if ($evento->coordenadorId == $usuarioLogado->id
-        || $usuariosDaComissaoCientifica->contains($usuarioLogado)
-        || $usuariosDaComissaoOrganizadora->contains($usuarioLogado)
-        || $evento->userIsCoordComissaoCientifica($usuarioLogado)
-        || $evento->userIsCoordComissaoOrganizadora($usuarioLogado)
-        || $trabalho->autorId == $usuarioLogado->id
-        || $trabalhosCoautor->contains($trabalho->id)
-        || $usuarioLogado->administradors()->exists()) {
+        if (
+            $evento->coordenadorId == $usuarioLogado->id
+            || $usuariosDaComissaoCientifica->contains($usuarioLogado)
+            || $usuariosDaComissaoOrganizadora->contains($usuarioLogado)
+            || $evento->userIsCoordComissaoCientifica($usuarioLogado)
+            || $evento->userIsCoordComissaoOrganizadora($usuarioLogado)
+            || $trabalho->autorId == $usuarioLogado->id
+            || $trabalhosCoautor->contains($trabalho->id)
+            || $usuarioLogado->administradors()->exists()
+        ) {
             // dd($arquivo);
             if ($arquivo != null && Storage::disk()->exists($arquivo->nome)) {
                 return Storage::download($arquivo->nome);
@@ -1063,7 +1071,7 @@ class TrabalhoController extends Controller
             ]);
         }
 
-        return redirect()->back()->with(['mensagem' => 'Correção de '.$trabalho->titulo.' enviada com sucesso!']);
+        return redirect()->back()->with(['mensagem' => 'Correção de ' . $trabalho->titulo . ' enviada com sucesso!']);
     }
 
     public function downloadArquivoCorrecao(Request $request)
@@ -1098,7 +1106,8 @@ class TrabalhoController extends Controller
                         return $trabalho->autor->name;
                     },
                     SORT_REGULAR,
-                    $direcao == $direction));
+                    $direcao == $direction
+                ));
             }
         } elseif ($column == 'titulo') {
             $trabalhos = collect();
@@ -1108,7 +1117,8 @@ class TrabalhoController extends Controller
                         return $trabalho->titulo;
                     },
                     SORT_REGULAR,
-                    $direcao == $direction));
+                    $direcao == $direction
+                ));
             }
         } elseif ($column == 'areaId') {
             $trabalhos = collect();
@@ -1118,7 +1128,8 @@ class TrabalhoController extends Controller
                         return $trabalho->area->nome;
                     },
                     SORT_REGULAR,
-                    $direcao == $direction));
+                    $direcao == $direction
+                ));
             }
         }
 
@@ -1156,7 +1167,8 @@ class TrabalhoController extends Controller
         $justificativa = str_replace(
             ['%NOME_AUTOR%', '%TITULO_TRABALHO%', '%NOME_EVENTO%', '%NOME_MODALIDADE%', '%NOME_AREA%'],
             [$trabalho->autor->name, $trabalho->titulo, $trabalho->evento->nome, $trabalho->modalidade->nome, $trabalho->area->nome],
-            $msgParecer->texto);
+            $msgParecer->texto
+        );
         Parecer::updateOrCreate(['parecer_final' => true, 'trabalhoId' => $trabalho->id], ['resultado' => $parecer, 'justificativa' => $justificativa]);
 
         $trabalho->update();
@@ -1207,7 +1219,7 @@ class TrabalhoController extends Controller
                 'nome' => $trab->autor->name,
                 'area' => $trab->area->nome,
                 'modalidade' => $trab->modalidade->nome,
-                'rota_download' => ! (empty($trab->arquivo->nome)) ? route('downloadTrabalho', ['id' => $trab->id]) : '#',
+                'rota_download' => !(empty($trab->arquivo->nome)) ? route('downloadTrabalho', ['id' => $trab->id]) : '#',
             ];
             $trabalhoJson->push($trabalho);
         }
@@ -1232,7 +1244,7 @@ class TrabalhoController extends Controller
         // dd($revisor);
         foreach ($modalidade->criterios as $criterio) {
             $validarCriterio = $request->validate([
-                'criterio_'.$criterio->id => 'required',
+                'criterio_' . $criterio->id => 'required',
             ]);
         }
 
@@ -1244,7 +1256,7 @@ class TrabalhoController extends Controller
         foreach ($modalidade->criterios as $criterio) {
             $avaliacao = new Avaliacao();
             $avaliacao->revisor_id = $revisor->id;
-            $avaliacao->opcao_criterio_id = $request->input('criterio_'.$criterio->id);
+            $avaliacao->opcao_criterio_id = $request->input('criterio_' . $criterio->id);
             $avaliacao->trabalho_id = $trabalho_id;
             $avaliacao->save();
         }
@@ -1307,7 +1319,7 @@ class TrabalhoController extends Controller
             }
 
             $extensao = $arquivo->getClientOriginalExtension();
-            if (! in_array($extensao, $tiposcadastrados)) {
+            if (!in_array($extensao, $tiposcadastrados)) {
                 return true;
             }
